Type Hardening, Consolidation, Assert Replacement
General:
chore: some code formatting
feat: changed guid definition to hoist the required type conversion, preventing the internal warning generation
feat: replaced macro std assert and macro assert usage with ::hg::assertd/rt
feat: fixed some template types by using underlying container type queries
feat: enforced typed constraints for type erase supported wrappers, preventing instantiation with incomplete types
feat: added explicit warning suppression to reduce warning noise when dealing with highly- and safely constrained type erase objects
feat: added overload struct helper to support type overloaded lambda expressions
feat: changed AutoArray<T, ...> to a concrete template type and added some utility function

+++ src/Engine.Assets/AssetGuid.cpp
@@ -8,20 +8,20 @@
 using namespace hg;
 
 asset_guid hg::generate_unstable_asset_guid() {
-    static std::atomic_uint_fast32_t ai { 1 };
+	static std::atomic_uint_fast32_t ai { 1 };
 
-    const auto pre = ai.fetch_add(1, std::memory_order_acq_rel);
+	const auto pre = ai.fetch_add(1, std::memory_order_acq_rel);
 
-    // TODO: Replace temporary solution
-    const auto tt = std::chrono::high_resolution_clock::now().time_since_epoch().count();
+	// TODO: Replace temporary solution
+	const auto tt = std::chrono::high_resolution_clock::now().time_since_epoch().count();
 
-    return { pre, 0, 0, tt };
+	return { pre, 0, 0, static_cast<u64>(tt) };
 }
 
 asset_guid hg::generate_asset_guid() {
 
-    asset_guid guid {};
-    GuidGenerate(guid);
+	asset_guid guid {};
+	GuidGenerate(guid);
 
-    return guid;
+	return guid;
 }
